# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IntensityResponse": {
        "properties": {
          "attribution": {
            "default": "Carbon intensity data: Energy-Charts.info, Fraunhofer ISE",
            "title": "Attribution",
            "type": "string"
          },
          "fetched_at": {
            "format": "date-time",
            "title": "Fetched At",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/PointOut"
            },
            "title": "Points",
            "type": "array"
          },
          "resolution_minutes": {
            "title": "Resolution Minutes",
            "type": "integer"
          },
          "zone": {
            "title": "Zone",
            "type": "string"
          }
        },
        "required": [
          "zone",
          "resolution_minutes",
          "fetched_at",
          "points"
        ],
        "title": "IntensityResponse",
        "type": "object"
      },
      "PointOut": {
        "properties": {
          "gco2_per_kwh": {
            "title": "Gco2 Per Kwh",
            "type": "number"
          },
          "is_forecast": {
            "title": "Is Forecast",
            "type": "boolean"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "gco2_per_kwh",
          "is_forecast"
        ],
        "title": "PointOut",
        "type": "object"
      },
      "ScheduleRequest": {
        "properties": {
          "deadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Latest time the job must finish. Defaults to the end of the forecast.",
            "title": "Deadline"
          },
          "duration_minutes": {
            "examples": [
              120
            ],
            "exclusiveMinimum": 0.0,
            "maximum": 2880.0,
            "title": "Duration Minutes",
            "type": "integer"
          },
          "power_kw": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100000.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Average power draw of the job. If set, emissions are returned in grams.",
            "examples": [
              0.4
            ],
            "title": "Power Kw"
          },
          "zones": {
            "description": "Zones the job may run in, in order of preference. The first one is used as the run-now baseline.",
            "examples": [
              [
                "de",
                "fr",
                "nl"
              ]
            ],
            "items": {
              "pattern": "^[A-Za-z]{2}$",
              "type": "string"
            },
            "maxItems": 10,
            "minItems": 1,
            "title": "Zones",
            "type": "array"
          }
        },
        "required": [
          "zones",
          "duration_minutes"
        ],
        "title": "ScheduleRequest",
        "type": "object"
      },
      "ScheduleResponse": {
        "properties": {
          "attribution": {
            "default": "Carbon intensity data: Energy-Charts.info, Fraunhofer ISE",
            "title": "Attribution",
            "type": "string"
          },
          "delay_minutes": {
            "title": "Delay Minutes",
            "type": "integer"
          },
          "emissions_saved_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emissions Saved G"
          },
          "recommended": {
            "$ref": "#/components/schemas/WindowOut"
          },
          "run_now": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WindowOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "Starting immediately in the first zone. Null if the data does not cover it."
          },
          "savings_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Savings Pct"
          }
        },
        "required": [
          "recommended",
          "run_now",
          "savings_pct",
          "emissions_saved_g",
          "delay_minutes"
        ],
        "title": "ScheduleResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WindowOut": {
        "properties": {
          "avg_gco2_per_kwh": {
            "title": "Avg Gco2 Per Kwh",
            "type": "number"
          },
          "emissions_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emissions G"
          },
          "end": {
            "format": "date-time",
            "title": "End",
            "type": "string"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          },
          "uses_forecast": {
            "title": "Uses Forecast",
            "type": "boolean"
          },
          "zone": {
            "title": "Zone",
            "type": "string"
          }
        },
        "required": [
          "zone",
          "start",
          "end",
          "avg_gco2_per_kwh",
          "uses_forecast"
        ],
        "title": "WindowOut",
        "type": "object"
      },
      "ZoneOut": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name"
        ],
        "title": "ZoneOut",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Carbon-aware scheduling for compute jobs on European power grids.",
    "title": "GridShift",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/healthz": {
      "get": {
        "operationId": "healthz_healthz_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Healthz Healthz Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Healthz",
        "tags": [
          "ops"
        ]
      }
    },
    "/v1/intensity/{zone}": {
      "get": {
        "description": "Measured and forecast carbon intensity (gCO2eq/kWh) for a zone.",
        "operationId": "get_intensity_v1_intensity__zone__get",
        "parameters": [
          {
            "description": "ISO 3166-1 alpha-2 country code, e.g. de",
            "in": "path",
            "name": "zone",
            "required": true,
            "schema": {
              "description": "ISO 3166-1 alpha-2 country code, e.g. de",
              "pattern": "^[A-Za-z]{2}$",
              "title": "Zone",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_past",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Past",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntensityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Intensity",
        "tags": [
          "intensity"
        ]
      }
    },
    "/v1/schedule": {
      "post": {
        "description": "Find the lowest-carbon start time and zone that still meets the deadline.",
        "operationId": "schedule_v1_schedule_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Schedule",
        "tags": [
          "schedule"
        ]
      }
    },
    "/v1/zones": {
      "get": {
        "description": "Countries with carbon intensity data, sorted by name.",
        "operationId": "zones_v1_zones_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ZoneOut"
                  },
                  "title": "Response Zones V1 Zones Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Zones",
        "tags": [
          "intensity"
        ]
      }
    }
  }
}